moved module files into src directory
#356
