+++ src/climate_indices/eto.py
@@ -28,7 +28,7 @@ import math
 import numba
 import numpy as np
 
-from climate_indices import utils
+from src.climate_indices import utils
 
 # ------------------------------------------------------------------------------
 # Retrieve logger and set desired logging level
