+++ src/climate_indices/palmer.py
@@ -6,7 +6,7 @@ import warnings
 import numba
 import numpy as np
 
-from climate_indices import utils
+from src.climate_indices import utils
 
 # ------------------------------------------------------------------------------
 # Retrieve logger and set desired logging level
